refactored entity to become DTO
