+++ src/main/java/dog/ticketlords/TicketlordsBE/DTO/VendorRatingDTO.java
@@ -1,0 +1,6 @@
+package dog.ticketlords.TicketlordsBE.DTO;
+
+import java.math.BigDecimal;
+
+public record VendorRatingDTO(String vendorName, BigDecimal avgVendorRating) {
+}
