--- src/main/java/dog/ticketlords/TicketlordsBE/entity/VendorRating.java
@@ -1,6 +1,0 @@
-package dog.ticketlords.TicketlordsBE.entity;
-
-import java.math.BigDecimal;
-
-public record VendorRating(String vendorName, BigDecimal avgVendorRating) {
-}
