+++ src/main/java/dog/ticketlords/TicketlordsBE/repositories/EventClicksRepository.java
@@ -8,14 +8,14 @@ import org.springframework.data.repository.query.Param;
 
 import dog.ticketlords.TicketlordsBE.dbentity.EventClicks;
 import dog.ticketlords.TicketlordsBE.dbentity.EventClicksId;
-import dog.ticketlords.TicketlordsBE.entity.CategoryClicksDTO;
+import dog.ticketlords.TicketlordsBE.DTO.CategoryClicksDTO;
 
 public interface EventClicksRepository extends JpaRepository<EventClicks, EventClicksId> {
   long countByEvent_EventId(long eventId);
 
   List<EventClicks> findAllByEvent_EventId(long eventId);
 
-  @Query("SELECT new dog.ticketlords.TicketlordsBE.entity.CategoryClicksDTO(c.categoryId, c.categoryName, COUNT(ec)) " +
+  @Query("SELECT new dog.ticketlords.TicketlordsBE.DTO.CategoryClicksDTO(c.categoryId, c.categoryName, COUNT(ec)) " +
       "FROM EventClicks ec " +
       "JOIN ec.event e " +
       "JOIN e.category c " +
